+++ pyproject.toml
@@ -87,6 +87,8 @@ lxml = {version = "^4.9.2", optional = true}
 pymupdf = {version = "^1.22.3", optional = true}
 pypdfium2 = {version = "^4.10.0", optional = true}
 gql = {version = "^3.4.1", optional = true}
+pandas = {version = "^2.0.1", optional = true}
+telethon = {version = "^1.28.5", optional = true}
 
 [tool.poetry.group.docs.dependencies]
 autodoc_pydantic = "^1.8.0"
@@ -190,6 +192,8 @@ extended_testing = [
  "pypdfium2",
  "tqdm",
  "lxml",
+ "pandas",
+ "telethon",
 ]
 
 [tool.ruff]
