+++ tests/unit_tests/document_loaders/blob_loaders/test_schema.py
