+++ tests/unit_tests/document_loaders/test_docs/telegram_channel.json
@@ -1,0 +1,34 @@
+[
+  {
+    "sender_id": -1111111,
+    "text": "Hello, world!",
+    "date": "2023-05-15T19:30:49+00:00",
+    "message.id": 1785,
+    "is_reply": false,
+    "reply_to_id": null
+  },
+  {
+    "sender_id": -1111111,
+    "text": "Telegram is the best!",
+    "date": "2023-05-08T20:17:10+00:00",
+    "message.id": 1784,
+    "is_reply": true,
+    "reply_to_id": 1783
+  },
+  {
+    "sender_id": -1111111,
+    "text": "Langchain is great.",
+    "date": "2023-05-03T23:43:33+00:00",
+    "message.id": 1783,
+    "is_reply": true,
+    "reply_to_id": 1782
+  },
+  {
+    "sender_id": -1111111,
+    "text": "LLMs are awesome!",
+    "date": "2023-05-03T15:32:25+00:00",
+    "message.id": 1782,
+    "is_reply": false,
+    "reply_to_id": null
+  }
+]
